Separate DynamoDB functionalities into separate file
